requirements.txt
flask
langchain
langchain-community
langchain-groq
sentence-transformers
python-dotenv
faiss-cpu